Split BitcoinRPC
