+++ src/checkpoints.cpp
@@ -42,9 +42,6 @@ namespace Checkpoints
 	( 1500000,	uint256("0xd9d438cf90d7fbc8367c904b5f70c0020c7a0465ecddaf53c4f885de08316d2d"))
 	( 1600000,	uint256("0x889b5172d5e89da5237ba634053b380446f65d4b221e4cc5c4a05b946f1784e6"))
 	( 1610644,	uint256("0x8db050db26b0f9177b0ad29329ef6840eb136dd2769e1e6844343939ca570796"))
-	( 1700000,	uint256("0x000000029707865bb9e7a4538157ad03b4516604d1c544669d7d03ada8ea74d7"))
-	( 1800000,	uint256("0x7cbc53ad650b56194229cf51d96643fcababa54ef284854181c3e869fbd701ab"))
-	( 1900000,	uint256("0x8d712fe472917cc98b03c6fbc52d5ab0c7e9f65f8ee5f15d2e5ea103427d4f09"))
 	;
 
     static MapCheckpoints mapCheckpointsTestnet =
