+++ src/makefile.bsd
@@ -127,7 +127,9 @@ OBJS= \
     obj/main.o \
     obj/net.o \
     obj/protocol.o \
-    obj/bitcoinrpc.o \
+    obj/rpcclient.o \
+    obj/rpcprotocol.o \
+    obj/rpcserver.o \
     obj/rpcdump.o \
     obj/rpcnet.o \
     obj/rpcmining.o \
