+++ src/makefile.osx
@@ -87,7 +87,9 @@ OBJS= \
     obj/main.o \
     obj/net.o \
     obj/protocol.o \
-    obj/bitcoinrpc.o \
+    obj/rpcclient.o \
+    obj/rpcprotocol.o \
+    obj/rpcserver.o \
     obj/rpcdump.o \
     obj/rpcnet.o \
     obj/rpcmining.o \
